fix rounding error in coordinatemapper

+++ frgpascal/hardware/geometry.py
@@ -34,7 +34,7 @@ class CoordinateMapper:
         if len(p) == 2:
             p = list(p)
             p.append(0)
-        p = np.asarray(p)
+        p = np.asarray(p, dtype=float)
         p[2] = self.zinterp(p[:2])
         pmap = p - self.xyoffset
 
@@ -205,7 +205,11 @@ class Workspace:
         self.gantry.moveto(*self.p0)
         self.gripper.open(self.OPENWIDTH)
         self.transform = map_coordinates(
-            self.name, self.testslots, self.testpoints, self.gantry, self.z_clearance,
+            self.name,
+            self.testslots,
+            self.testpoints,
+            self.gantry,
+            self.z_clearance,
         )
         self.__calibrated = True
 
@@ -301,6 +305,7 @@ class Workspace:
         plt.legend(bbox_to_anchor=(1.05, 1), loc=2, borderaxespad=0.0)
         plt.title(tray.name)
         plt.yticks(
-            yvals[::-1], [chr(65 + i) for i in range(len(yvals))],
+            yvals[::-1],
+            [chr(65 + i) for i in range(len(yvals))],
         )
         plt.xticks(xvals, [i + 1 for i in range(len(xvals))])
